second commit

+++ internal/types/token.go
@@ -4,6 +4,6 @@ import "github.com/dgrijalva/jwt-go"
 
 type TokenClaims struct {
 	jwt.StandardClaims
-	User_id  int64  `json:"user_id"`
+	User_id  int    `json:"user_id"`
 	Username string `json:"username"`
 }
